Random Blind Search
Blindly and randomly tries to search between two random Wikipedia
articles. There was no way this was going to work, but we had to give
it a try!

init.json is a data file that provides a start and end point for the
chain.

More detailed notes can be found on my website
(http://rlamacraft.co.uk/projects/WikiMind#RandomBlindSearch)