Improved learned score
Calculate the score, to commit to the machine learning process, to take into account the length of the chain the link was found in whilst maintaining the bounds of the possible score classes (currently 0 to 100) with each iteration. In other words, the possible values outputted by the classifier should not increase with each round of the learning process.